+++ pkg/sass-parser/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 0.4.9-dev
+## 0.4.9
 
 * No user-visible changes.