+++ pkg/sass_api/CHANGELOG.md
@@ -1,4 +1,4 @@
-## 15.0.1-dev
+## 15.0.1
 
 * No user-visible changes.